ml-service/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
numpy==1.22.4
spacy==3.0.6
requests==2.29.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pydantic==1.7.4
